backend/requirements.txt
# Web Framework
fastapi
uvicorn
pydantic

# HTTP and Network
requests

# Time and Scheduling
pytz
apscheduler
schedule
chinesecalendar

# AI Services
autoagents-core

# Data Processing and Scraping
beautifulsoup4
lxml
playwright
tqdm

# Database
supabase

# Configuration
PyYAML
python-dotenv

# Logging
loguru